+++ server/prisma/migrations/20260708073927_init/migration.sql
@@ -1,0 +1,7 @@
+-- CreateTable
+CREATE TABLE "Organization" (
+    "id" TEXT NOT NULL PRIMARY KEY,
+    "name" TEXT NOT NULL,
+    "type" TEXT NOT NULL,
+    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
+);